page routes

+++ controllers/section.js
@@ -14,8 +14,6 @@ exports.createSection = (req, res, next) => {
 exports.editOneSection = (req, res, next) => {
     // THE IMAGE IS ALREADY UPLOADED WHEN WE ARRIVE HERE
 
-    // WHAT IS IN REQ IF I USE 2 MULTERS ?
-
     let sectionReq = req.file ?
         {
             ...JSON.parse(req.body.section),
@@ -81,25 +79,33 @@ exports.deleteOneSection = (req, res, next) => {
     // WE DELETE THE ID IN ALL ARRAY SECTIONSIDS OF THE GROUPSECTIONS (like a cascade in php)
     // https://dev.to/kwabenberko/implementing-sql--like-cascades-in-mongoose-bap
     GroupSection.find({ sectionsIds: { $in: [sectionId] } }).then(groupSections => {
-        Promise.all(
-            groupSections.map(groupSection =>
-                GroupSection.findOneAndUpdate(
-                    { _id: groupSection._id },
-                    { $pull: { sectionsIds: sectionId } },
-                    { new: true }
+        if (groupSections.length > 0) {
+            Promise.all(
+                groupSections.map(groupSection =>
+                    GroupSection.findOneAndUpdate(
+                        { _id: groupSection._id },
+                        { $pull: { sectionsIds: sectionId } },
+                        { new: true }
+                    )
                 )
-            )
-        );
+            );
+        }
     });
 
     Section.findOne({ _id: sectionId })
         .then(section => {
-            const filename = section.mainImgUrl.split('/images/sections/')[1];
-            fs.unlink(`images/${filename}`, () => {
+            if (section.mainImgUrl) {
+                const filename = section.mainImgUrl.split('/images/sections/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Section.deleteOne({ _id: sectionId })
+                        .then(() => res.status(200).json({ message: 'Section supprimée !' }))
+                        .catch(error => res.status(400).json({ error }));
+                });
+            } else {
                 Section.deleteOne({ _id: sectionId })
                     .then(() => res.status(200).json({ message: 'Section supprimée !' }))
                     .catch(error => res.status(400).json({ error }));
-            });
+            }
         })
         .catch(error => res.status(500).json({ error }));
 
@@ -142,7 +148,6 @@ exports.getSeveralSections = (req, res, next) => {
 };
 
 exports.getOneSectionWithTitle = (req, res, next) => {
-    let sectionReq = req.body;
     Section.findOne({ title: String(req.body.title) }).then(
         (section) => {
             res.status(200).json(section);
